requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
dj-database-url==2.0.0
Django==4.2.18
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fonttools==4.49.0
frozenlist==1.4.1
googlemaps==4.10.0
html5lib==1.1
http_ece==1.2.0
hyperlink==21.0.0
idna==3.4
importlib_resources==6.4.5
incremental==24.7.2
inflection==0.5.1
isodate==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.24.4
openpyxl==3.1.2
packaging==24.2
pandas==2.0.3
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
py-vapid==1.9.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.2.1
pyphen==0.14.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pywebpush==2.0.0
PyYAML==6.0.2
redis==5.0.0
referencing==0.35.1
reportlab==4.0.4
requests==2.31.0
rpds-py==0.20.1
secure-smtplib==0.1.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
validate-email-address==1
validate_email==1.3
vine==5.1.0
wcwidth==0.2.6
weasyprint==61.2
webencodings==0.5.1
yarl==1.9.4
zipp==3.20.2
zope.interface==6.4.post2
zopfli==0.2.3
